Updated api results module and import.
